feat: request & transmit status packet; display proper info on Page::Info

+++ src/common/packets.hpp
@@ -53,7 +53,7 @@ struct StatusPacket
 		uint8_t flags;
 	};
 	uint8_t signalRating;
-	uint16_t battery;
+	float battery;
 };
 
 struct ReceiverSignal
